## Socialite or Guzzle making
i think guzzle is what they want.. so lets see how it goes

+++ server/app/Services/Implement/CommentServiceImp.php
@@ -34,6 +34,7 @@ class CommentServiceImp implements CommentService
     {
         $comment = $this->comment_repository->getOne($content['comment_id']);
         if (!$comment->id) throw new \App\Exceptions\ModuleNotFound('comment not Found');
+        if (strcmp($comment->post_id,$content['post_id'])) throw new \App\Exceptions\ModuleNotFound('Post not Found');
         if (!$comment->user_id) throw new IllegalUserApproach();
         $comment = $this->comment_repository->updateByContent($content);
         return $comment;
@@ -43,6 +44,7 @@ class CommentServiceImp implements CommentService
     {
         $comment = $this->comment_repository->getOne($content['comment_id']);
         if (!$comment->id) throw new \App\Exceptions\ModuleNotFound('comment not Found');
+        if (strcmp($comment->post_id,$content['post_id'])) throw new \App\Exceptions\ModuleNotFound('Post not Found');
         if (!$comment->user_id) throw new IllegalUserApproach();
         $result = $this->comment_repository->delete($comment);
         if (!$result) throw new \App\Exceptions\ModuleNotFound('delete failed');
